Implement cache busting favicons via 11ty image
Added a `favicon` shortcode that generates a hashed version of a favicon
and a corresponding `<link>` tag.

Currently this works for `icon.svg` and `apple-touch-icon.png`. I moved
them out of the `public` directory into `content/favicon` to prevent
generating two versions of the same icon - e.g., `icon.svg` and its
hashed version - in the output directory.